Initial Commit
This repo will have the simulation parameters used to generate the tracks used
for my Tracking paper and my dissertation.

In addition, it has the parameter space used to find optimal tracker parameter
values to use for further analyzes.

A later commit will include a Makefile to ease the creation and analysis of
these simulations.